// src/api/models/AccountTreeDataDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AccountDto } from "./AccountDto";

export type AccountTreeDataDto = {
  assets: Array<AccountDto>;
  liabilities: Array<AccountDto>;
  expense: Array<AccountDto>;
  income: Array<AccountDto>;
};